[refactor] signIn 로직 수정
- 네이밍 변경 : logIn -> signIn
- LoginService(inf) 삭제
- checkSigninInfo 메서드 생
