--- src/main/java/flab/just10minutes/member/service/LoginService.java
@@ -1,9 +1,0 @@
-package flab.just10minutes.member.service;
-
-import flab.just10minutes.member.dto.LoginRequest;
-
-public interface LoginService {
-    void logIn(Long uniqueId);
-    void logOut();
-    void loginValidate();
-}
